Fix to httpd_filter_name()
httpd_filter_name() was changed. The former code used sprintf format "%c" to append a single character. The new code uses pointer arithmetic to append a single character. However, sprintf also appends a NULL terminator to terminate the string, the new code does not. The top of the match code uses strlen, which means, that if a terminator is not added, bad things happen.